requirements.txt
pyserial
termcolor
MySQLdb